+++ index.android.js
@@ -1,4 +1,4 @@
 import { AppRegistry } from 'react-native'
-import setup from './js/pages/AppRouter'
+import setup from './js/AppRouter'
 
 AppRegistry.registerComponent('oneLibrary', () => setup)
