+++ index.ios.js
@@ -1,5 +1,5 @@
 import { AppRegistry } from 'react-native'
-import AppRouter from './js/pages/AppRouter'
+import AppRouter from './js/AppRouter'
 
 
 AppRegistry.registerComponent('oneLibrary', () => AppRouter)
