+++ js/AppRouter.js
@@ -1,19 +1,19 @@
 import React, { Component } from 'React'
-import LoginPage from './Login'
-import HomePage from './Home'
-import SearchPage from './Search'
-import BookInfo from './BookInfo'
-import BookCollect from './BookCollect'
-import BookCollectAdd from './BookCollectAdd'
-import BookCollectList from './BookCollectList'
-import RentInfo from './RentInfo'
-import Setting from './Setting'
-import FeedBack from './FeedBack'
-import SearchResult from './SearchResult'
-import AboutUs from './AboutUs'
-import AboutUsWeb from './AboutUsWeb'
+import LoginPage from './pages/Login'
+import HomePage from './pages/Home'
+import SearchPage from './pages/Search'
+import BookInfo from './pages/BookInfo'
+import BookCollect from './pages/BookCollect'
+import BookCollectAdd from './pages/BookCollectAdd'
+import BookCollectList from './pages/BookCollectList'
+import RentInfo from './pages/RentInfo'
+import Setting from './pages/Setting'
+import FeedBack from './pages/FeedBack'
+import SearchResult from './pages/SearchResult'
+import AboutUs from './pages/AboutUs'
+import AboutUsWeb from './pages/AboutUsWeb'
 import { Scene, Router, ActionConst, Actions } from 'react-native-router-flux'
-import * as scenes from '../constants/scene'
+import * as scenes from './constants/scene'
 
 export default class AppRouter extends Component {
   render() {
